+++ src/Controller/Admin/ReportCrudController.php
@@ -12,8 +12,10 @@ use EasyCorp\Bundle\EasyAdminBundle\Field\BooleanField;
 use EasyCorp\Bundle\EasyAdminBundle\Field\CollectionField;
 use EasyCorp\Bundle\EasyAdminBundle\Field\DateTimeField;
 use EasyCorp\Bundle\EasyAdminBundle\Field\IdField;
+use EasyCorp\Bundle\EasyAdminBundle\Field\TextEditorField;
 use EasyCorp\Bundle\EasyAdminBundle\Field\TextField;
 use EasyCorp\Bundle\EasyAdminBundle\Field\UrlField;
+use Symfony\Component\Translation\TranslatableMessage;
 
 class ReportCrudController extends AbstractCrudController
 {
@@ -36,68 +38,87 @@ class ReportCrudController extends AbstractCrudController
     #[\Override]
     public function configureFields(string $pageName): iterable
     {
-        $title = TextField::new('sys_title')->setLabel('entity.report.sys_title');
-        $title2 = $title->setFormTypeOption('disabled', 'disabled');
-        $internalId = IdField::new('sys_internal_id')->setLabel('entity.report.sys_internal_id');
-        $edocUrl = UrlField::new('edoc_url')->setLabel('entity.report.edoc_url');
-        $sys_link = UrlField::new('sys_link')->setLabel('entity.report.sys_link');
-        $name = TextField::new('name')->setLabel('entity.report.name');
-        $text = TextField::new('text')->setLabel('entity.report.text');
-        $systemOwner = TextField::new('sys_system_owner')->setLabel('entity.report.sys_system_owner');
-        $groups = AssociationField::new('groups')->setLabel('entity.report.groups');
+        switch ($pageName) {
+            case Crud::PAGE_INDEX:
+                // Cf. https://github.com/itk-dev/sysstatus/blob/5383a3a566ce316c338441ed826ecf3fdcf98815/src/Controller/AdminController.php#L263-L288
+                yield IdField::new('id')->setLabel('entity.report.sys_id');
+                yield TextField::new('sysTitle')->setLabel('entity.report.sys_title');
+                yield CollectionField::new('groups')->setLabel('entity.report.groups')
+                    ->renderExpanded();
+                yield TextField::new('sysOwnerSub')->setLabel('entity.report.sys_owner_sub');
+                yield TextField::new('sysSystemOwner')->setLabel('entity.report.sys_system_owner');
+                yield UrlField::new('sysLink')->setLabel('entity.report.sys_id')->setLabel('entity.report.sys_link')
+                    ->formatValue(static fn ($value) => new TranslatableMessage('Link'));
+                yield BooleanField::new('textSet')->setLabel('entity.report.text')
+                    ->renderAsSwitch(false);
 
-        $answerarea = CollectionField::new('answerarea')->setTemplatePath('easy_admin_overrides/answers_show.html.twig')->setLabel('Smileys');
-        $answerarea = CollectionField::new('answerarea')->setTemplatePath('easy_admin_overrides/answers_show.html.twig')->setLabel('Smileys');
+                return;
 
-        $alternativeTitle = TextField::new('sys_alternative_title')->setLabel('entity.report.sys_alternative_title');
-        $sys_updated = DateTimeField::new('sys_updated')->setLabel('entity.report.sys_updated');
-        $owner = TextField::new('sys_owner')->setLabel('entity.report.sys_owner');
-        $afdeling = TextField::new('sys_owner_sub')->setLabel('entity.report.sys_owner');
-        $confidentialInformation = BooleanField::new('sys_confidential_information')->setLabel('entity.report.sys_confidential_information');
-        $purpose = TextField::new('sys_purpose')->setLabel('entity.report.sys_purpose');
-        $classification = TextField::new('sys_classification')->setLabel('entity.report.sys_classification');
-        $dateForRevision = DateTimeField::new('sys_date_for_revision')->setLabel('entity.report.sys_date_for_revision');
-        $persons = TextField::new('sys_persons')->setLabel('entity.report.sys_persons');
-        $informationTypes = TextField::new('sys_information_types')->setLabel('entity.report.sys_information_types');
-        $dataSentTo = TextField::new('sys_data_sent_to')->setLabel('entity.report.sys_data_sent_to');
-        $dataComeFrom = TextField::new('sys_data_come_from')->setLabel('entity.report.sys_data_come_from');
-        $dataLocation = TextField::new('sys_data_location')->setLabel('entity.report.sys_data_location');
-        $deletionDate = TextField::new('sys_latest_deletion_date')->setLabel('entity.report.sys_latest_deletion_date');
-        $dataWorthSaving = TextField::new('sys_data_worth_saving')->setLabel('entity.report.sys_data_worth_saving');
-        $dataProcessors = TextField::new('sys_data_processors')->setLabel('entity.report.sys_data_processors');
-        $dataProcessingAgreement = TextField::new('sys_data_processing_agreement')->setLabel('entity.report.sys_data_processing_agreement');
-        $dataProcessingAgreementLink = TextField::new('sys_data_processing_agreement_link')->setLabel('entity.report.sys_data_processing_agreement_link');
-        $auditorStatement = TextField::new('sys_auditor_statement')->setLabel('entity.report.sys_auditor_statement');
-        $auditorStatementLink = TextField::new('sys_auditor_statement_link')->setLabel('entity.report.sys_auditor_statement_link');
-        $dataToScience = TextField::new('sys_data_to_science')->setLabel('entity.report.sys_data_to_science');
-        $usage = TextField::new('sys_usage')->setLabel('entity.report.sys_usage');
-        $requestForInsight = TextField::new('sys_request_for_insight')->setLabel('entity.report.sys_request_for_insight');
-        $dateUse = DateTimeField::new('sys_date_use')->setLabel('entity.report.sys_date_use');
-        $status = TextField::new('sys_status')->setLabel('entity.report.sys_status');
-        $remarks = TextField::new('sys_remarks')->setLabel('entity.report.sys_remarks');
-        $internalInformation = TextField::new('sys_internal_information')->setLabel('entity.report.sys_internal_information');
-        $obligationToInform = TextField::new('sys_obligation_to_inform')->setLabel('entity.report.sys_obligation_to_inform');
-        $legalBasis = TextField::new('sys_legal_basis')->setLabel('entity.report.sys_legal_basis');
-        $consent = TextField::new('sys_consent')->setLabel('entity.report.sys_consent');
-        $impactAnalysis = TextField::new('sys_impact_analysis')->setLabel('entity.report.sys_impact_analysis');
-        $impactAnalysisLink = TextField::new('sys_impact_analysis_link')->setLabel('entity.report.sys_impact_analysis_link');
-        $authorizationProcedure = TextField::new('sys_authorization_procedure')->setLabel('entity.report.sys_authorization_procedure');
-        $version = TextField::new('sys_version')->setLabel('entity.report.sys_version');
+            case Crud::PAGE_DETAIL:
+                // Cf. https://github.com/itk-dev/sysstatus/blob/5383a3a566ce316c338441ed826ecf3fdcf98815/config/packages/easy_admin.yaml#L100-L143
+                yield TextField::new('sysTitle')->setLabel('entity.report.sys_title');
+                yield IdField::new('sysInternalId')->setLabel('entity.report.sys_internal_id');
+                yield UrlField::new('eDocUrl')->setLabel('entity.report.edoc_url');
+                yield UrlField::new('sysLink')->setLabel('entity.report.sys_link');
+                yield TextField::new('name')->setLabel('entity.report.name');
+                yield TextEditorField::new('text')->setLabel('entity.report.text')
+                    // Show raw value
+                    ->setTemplatePath('admin/text_editor.raw.html.twig')
+                ;
+                yield TextField::new('sys_system_owner')->setLabel('entity.report.sys_system_owner');
+                // @todo Add links to each group?
+                yield CollectionField::new('groups')->setLabel('entity.report.groups');
+                yield CollectionField::new('answerarea')->setTemplatePath('easy_admin_overrides/answers_show.html.twig')->setLabel('entity.report.answers');
+                yield TextField::new('sys_alternative_title')->setLabel('entity.report.sys_alternative_title');
+                yield DateTimeField::new('sys_updated')->setLabel('entity.report.sys_updated');
+                yield TextField::new('sys_owner')->setLabel('entity.report.sys_owner');
+                yield BooleanField::new('sys_confidential_information')->setLabel('entity.report.sys_confidential_information');
+                yield TextField::new('sys_purpose')->setLabel('entity.report.sys_purpose');
+                yield TextField::new('sys_classification')->setLabel('entity.report.sys_classification');
+                yield DateTimeField::new('sys_date_for_revision')->setLabel('entity.report.sys_date_for_revision');
+                yield TextField::new('sys_persons')->setLabel('entity.report.sys_persons');
+                yield TextField::new('sys_information_types')->setLabel('entity.report.sys_information_types');
+                yield TextField::new('sys_data_sent_to')->setLabel('entity.report.sys_data_sent_to');
+                yield TextField::new('sys_data_come_from')->setLabel('entity.report.sys_data_come_from');
+                yield TextField::new('sys_data_location')->setLabel('entity.report.sys_data_location');
+                yield TextField::new('sys_latest_deletion_date')->setLabel('entity.report.sys_latest_deletion_date');
+                yield TextField::new('sys_data_worth_saving')->setLabel('entity.report.sys_data_worth_saving');
+                yield TextField::new('sys_data_processors')->setLabel('entity.report.sys_data_processors');
+                yield TextField::new('sys_data_processing_agreement')->setLabel('entity.report.sys_data_processing_agreement');
+                yield TextField::new('sys_data_processing_agreement_link')->setLabel('entity.report.sys_data_processing_agreement_link');
+                yield TextField::new('sys_auditor_statement')->setLabel('entity.report.sys_auditor_statement');
+                yield TextField::new('sys_auditor_statement_link')->setLabel('entity.report.sys_auditor_statement_link');
+                yield TextField::new('sys_data_to_science')->setLabel('entity.report.sys_data_to_science');
+                yield TextField::new('sys_usage')->setLabel('entity.report.sys_usage');
+                yield TextField::new('sys_request_for_insight')->setLabel('entity.report.sys_request_for_insight');
+                yield DateTimeField::new('sys_date_use')->setLabel('entity.report.sys_date_use');
+                yield TextField::new('sys_status')->setLabel('entity.report.sys_status');
+                yield TextField::new('sys_remarks')->setLabel('entity.report.sys_remarks');
+                yield TextField::new('sys_internal_information')->setLabel('entity.report.sys_internal_information');
+                yield TextField::new('sys_obligation_to_inform')->setLabel('entity.report.sys_obligation_to_inform');
+                yield TextField::new('sys_legal_basis')->setLabel('entity.report.sys_legal_basis');
+                yield TextField::new('sys_consent')->setLabel('entity.report.sys_consent');
+                yield TextField::new('sys_impact_analysis')->setLabel('entity.report.sys_impact_analysis');
+                yield TextField::new('sys_impact_analysis_link')->setLabel('entity.report.sys_impact_analysis_link');
+                yield TextField::new('sys_authorization_procedure')->setLabel('entity.report.sys_authorization_procedure');
+                yield TextField::new('sys_version')->setLabel('entity.report.sys_version');
 
-        if (Crud::PAGE_INDEX === $pageName) {
-            return [$title, $name, $systemOwner, $afdeling, $edocUrl, $text,  $sys_link,  $text,  $groups];
-        } elseif (Crud::PAGE_DETAIL === $pageName) {
-            return [$title, $internalId, $edocUrl,
-                $name, $text, $systemOwner, $groups, $answerarea, $alternativeTitle, $sys_updated, $owner,
-                $confidentialInformation, $purpose, $classification, $dateForRevision, $persons, $informationTypes,
-                $dataSentTo, $dataComeFrom, $dataLocation, $deletionDate, $dataWorthSaving, $dataProcessors,
-                $dataProcessingAgreement, $dataProcessingAgreementLink, $auditorStatement, $auditorStatementLink,
-                $dataToScience, $usage, $requestForInsight, $dateUse, $status, $remarks, $internalInformation, $obligationToInform,
-                $legalBasis, $consent, $impactAnalysis, $impactAnalysisLink, $authorizationProcedure, $version];
-        } elseif (Crud::PAGE_EDIT === $pageName) {
-            return [$title2, $edocUrl, $groups, $text];
-        } else {
-            throw new \Exception('Invalid page: '.$pageName);
+                return;
+
+            case Crud::PAGE_NEW:
+            case Crud::PAGE_EDIT:
+                // Cf. https://github.com/itk-dev/sysstatus/blob/5383a3a566ce316c338441ed826ecf3fdcf98815/config/packages/easy_admin.yaml#L144-L150
+                yield TextField::new('sysTitle')->setLabel('entity.report.sys_title')
+                    ->setDisabled();
+                yield TextEditorField::new('text');
+                yield AssociationField::new('groups')->setLabel('entity.report.groups')
+                    ->setPermission('ROLE_ADMIN');
+                yield UrlField::new('eDocUrl')->setLabel('entity.report.edoc_url');
+
+                return;
+
+            default:
+                throw new \Exception('Invalid page: '.$pageName);
         }
     }
 }