+++ CHANGELOG.md
@@ -8,6 +8,8 @@ See [keep a changelog] for information about writing changes to this log.
 
 ## [Unreleased]
 
+* [PR-32](https://github.com/itk-dev/sysstatus/pull/32)
+  3644: Cleaned up template and controller
 * [PR-31](https://github.com/itk-dev/sysstatus/pull/31)
   Upgraded to Symfony 7.2
 * [PR-30](https://github.com/itk-dev/sysstatus/pull/30)